extract icon-related background code to icon.js
